docker-requirements.txt
bosdyn-api==3.0.0
bosdyn-client==3.0.0
bosdyn-core==3.0.0
certifi==2020.12.5
chardet==4.0.0
Deprecated==1.2.10
grpcio==1.19.0
idna==2.10
numpy==1.19.5
Pillow==6.1.0
protobuf==3.7.1
PyJWT==2.0.1
requests==2.25.1
six==1.12.0
urllib3==1.26.2
wrapt==1.12.1
flask
